Subtitle-download/requirements.txt
python-telegram-bot==20.6
aiohttp==3.8.5
httpx~=0.25.0